tests/wpt/web-platform-tests/html/links/manifest/link-relationship/link-rel-manifest.html

<!DOCTYPE html>
<title>
  Test that "manifest" is a supported value for the `rel` of a `link`
</title>
<link rel="help" href="https://html.spec.whatwg.org/#link-type-manifest" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(() => {
    const result = document.createElement("link").relList.supports("manifest");
    assert_true(
      result,
      "Expected true if manifest is supported as a link relationship"
    );
  }, 'link element supports a rel value of "manifest".');
</script>
